requirements.txt
openai==1.11.1
streamlit==1.22.0
faiss-cpu==1.7.3
pandas==1.5.3
numpy==1.24.3
python-docx==0.8.11
python-dotenv==1.0.0
